Allow target_state to be None
As discussed in #24. This part of the the proposal should be
uncontroversial, and definitely needed for the implementation of the PE
optimization.

As a side-effect, `Objective` now only accepts keyword-arguments, not
positional arguments. This makes user-code robust against future changes
in the API, and it is a good idea in any case. Accidentally mixing up
the initial state and target states, or states and Hamiltonians could
happen very easily with positional arguments and lead to strange bugs
that are hard to figure out.