Number rst files and notebooks
If there is a single "Reference" page for the entire documentation, the
BibTeX plugin numbers citations in the order it processes them. This
order is alphabetially according to the file name of the rst file, not
the order in which the rst file are listed in `index.rst`. The only
way to make the two match is to number the rst files.

This also numbers the example notebooks, so they appear in the intended
order.