Remove PulseOptions in favor of simple dict
There is no need to use a specialized PulseOptions object, when a simple
dictionary will do. Also, it gives us significantly more flexibility.
I'm considering implementing a function that takes the same input as
optimize_pulses, but generates the necessary input data for a running
`qdyn_optimize`. In that case, the `pulse_options` might a good place to
store additional QDYN-specific config-file attributes.