Test only the first three notebooks in AppVeyor
AppVeyor seems pretty broken, and just hangs on a lot of the notebooks.
I'm not able to reproduce this when running on Windows locally. So
there's nothing really we can do about that: AppVeyor just doesn't
provide a great testing environment. As a compromise, we'll run the
first three notebooks through AppVeyor, so that we have a test for the
basics, at least. The more advanced examples will be tested on Travis
only (and manually, on Windows)