Allow to specific precision for F_avg
The imaginary part of F_avg (which mathematically should be zero) turned
out to be a bit larger than expected in practical applications. We don't
really want to crash irrecoverably in this case, but it's something a
user needs to keep an eye on. Therefore, the limit has to be
configurable.