Avoid using QuTiP 4.4.0 on Travis/Appveyor
All problems introduced in QuTiP 4.4 were worked around, except for
https://github.com/qutip/qutip/issues/1047

A fix to https://github.com/qutip/qutip/issues/1047 was merged into
QuTiP's master, so presumably the problem will disappear with the next
release of QuTiP.

This commit specifically excludes QuTiP 4.4.0 from being installed on
Travis and AppVeyor. We do not exclude that version in the list of
dependencies in setupy.py, because the issue is mostly cosmetic
(although it will fail tests). The average user should be OK using
krotov with QuTiP 4.4.0

Closes #49