Show an alert on the 'latest' docs on RTD
People have gotten confused by going to the 'latest' version of the
documentation and the documentation not matching the version of the
package that they installed with pip. Thus, we inject some javascript
that shows a warning that 'latest' is for an unreleased development
version.

This works in conjunction with a builtin feature on RTD to show warnings
for old version. This has to be enable in the RTD project settings
(Admin > Advanced Settings > Show version warning).