Don't print out relative phase in example 5
The relative phase that was being printed in the
`05_example_transmon_xgate.ipynb` was not numerically meaningful: The
ode-solver that is being used my `qutip.mesolve` is only accurate to
about 1e-7, and the (very small) relative phase that was being printed
was dominated by this error. QuTiP 4.4 made some changes to mesolve that
(withing the error of 1e-7) produces different results, and results in a
different relative phase in this example (even if printed to just one
significant digit).

On top of that, looking at the relative phase like we were doing is
actually not necessary: We already know that the phase is correct from
the reached value of J_T_re. There is no problem with looking at
population dynamics only of top of the J_T_re value.