Limit 'make test' to Python 3.7
Our tests take quite a long time, and it's not reasonable for the
default local testing during development ('make test') to test against
all versions of Python. Therefore, 'make test' now is the same as 'make
test37', and testing against other versions should be done manually
(e.g., 'make test36' or to the CI on Travis.