Use Gh-Releases for documentation artifacts
Adding a pdf of the documentation to the repo for every release is not
sustainable: These binary files are relatively large and will eventually
clog the repository.

I have not been able to to compile a completely satisfactory PDF
documentation on Travis: there seems to be some problems with getting
the unicode-enable fonts that we need without also messing up some of
the math rendering on Linux. Therefore, I'll stick to producing the
documentation artifacts locally during the release process and uploading
them manually to Github Releases.

The release script and documentation has been updated accordingly. This
also switches to Doctr-Versions-Menu for rendering the versions menu in
the documentation.